Extract auth status service
